Got expiration sorted out

+++ paniko/config/config.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Config struct {
-	DaysToExpiry int64
+	DaysToExpiry int
 	Mail         Mail
 }
